Fixed two bugs about button
