+++ react-app/src/Home/ButtonInCard.js
@@ -1,4 +1,4 @@
-import { StyleSheet, Pressable } from 'react-native'
+import { StyleSheet, Pressable, Alert } from 'react-native'
 import React from 'react'
 import { useNavigation } from '@react-navigation/native'
 import ArrowRightIcon from '../../assets/icons/arrow-right.svg'
@@ -9,7 +9,13 @@ export default function ButtonInCard({ available, screen }) {
   return (
     <Pressable
       style={[styles.button, !available && styles.unavailable]}
-      onPress={() => navigation.navigate(screen)}
+      onPress={
+        available
+          ? () => navigation.navigate(screen)
+          : () => {
+              Alert.alert('Coming soon!')
+            }
+      }
     >
       <ArrowRightIcon height={20} width={20} />
     </Pressable>
